+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/state/event/handler/internal/utils/SelfUserUtils.kt
@@ -38,4 +38,5 @@ internal fun MutableGlobalState.updateCurrentUser(currentUser: User?, receivedUs
     setTotalUnreadCount(me.totalUnreadCount)
     setChannelUnreadCount(me.unreadChannels)
     setUnreadThreadsCount(me.unreadThreads)
+    setBlockedUserIds(me.blockedUserIds)
 }